Flytt Korriger Vedtak-modal ut i egen fil
